+++ packages/rich-text/src/plugins/shared/EmbeddedBlockToolbarIcon.tsx
@@ -1,8 +1,10 @@
 import * as React from 'react';
 
 import { FieldAppSDK } from '@contentful/app-sdk';
-import { Flex, Icon, Menu } from '@contentful/f36-components';
-import { AssetIcon, EmbeddedEntryBlockIcon } from '@contentful/f36-icons';
+import { Flex, Menu } from '@contentful/f36-components';
+import { Icon } from '@contentful/f36-components';
+import { ImageSquareIcon, EmbeddedBlockIcon } from '@contentful/f36-icons';
+import tokens from '@contentful/f36-tokens';
 import { BLOCKS } from '@contentful/rich-text-types';
 import { css } from 'emotion';
 
@@ -48,24 +50,26 @@ export function EmbeddedBlockToolbarIcon({
   const type = getEntityTypeFromNodeType(nodeType);
   const baseClass = `rich-text__${nodeType}`;
   return (
-    <Menu.Item
-      disabled={isDisabled}
-      className={`${baseClass}-list-item`}
-      onClick={handleClick}
-      testId={`toolbar-toggle-${nodeType}`}
-    >
-      <Flex alignItems="center" flexDirection="row">
-        <Icon
-          as={type === 'Asset' ? AssetIcon : EmbeddedEntryBlockIcon}
-          className={`rich-text__embedded-entry-list-icon ${styles.icon}`}
-          variant="secondary"
-        />
-        <span>
-          {type}
-          {nodeType == BLOCKS.EMBEDDED_RESOURCE && <ResourceNewBadge />}
-        </span>
-      </Flex>
-    </Menu.Item>
+    <Menu>
+      <Menu.Item
+        disabled={isDisabled}
+        className={`${baseClass}-list-item`}
+        onClick={handleClick}
+        testId={`toolbar-toggle-${nodeType}`}
+      >
+        <Flex alignItems="center" flexDirection="row">
+          <Icon
+            as={type === 'Asset' ? ImageSquareIcon : EmbeddedBlockIcon}
+            className={`rich-text__embedded-entry-list-icon ${styles.icon}`}
+            color={tokens.gray900}
+          />
+          <span>
+            {type}
+            {nodeType == BLOCKS.EMBEDDED_RESOURCE && <ResourceNewBadge />}
+          </span>
+        </Flex>
+      </Menu.Item>
+    </Menu>
   );
 }