+++ packages/boolean/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-boolean@1.7.1...@contentful/field-editor-boolean@1.7.2) (2025-09-08)
+
+**Note:** Version bump only for package @contentful/field-editor-boolean
+
 ## [1.7.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-boolean@1.7.0...@contentful/field-editor-boolean@1.7.1) (2025-09-05)
 
 **Note:** Version bump only for package @contentful/field-editor-boolean