+++ packages/reference/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.10.8](https://github.com/contentful/field-editors/compare/@contentful/field-editor-reference@6.10.7...@contentful/field-editor-reference@6.10.8) (2025-09-08)
+
+**Note:** Version bump only for package @contentful/field-editor-reference
+
 ## [6.10.7](https://github.com/contentful/field-editors/compare/@contentful/field-editor-reference@6.10.6...@contentful/field-editor-reference@6.10.7) (2025-09-05)
 
 ### Bug Fixes